fix: refactor recommendation service (2/2)
