+++ src/spaceone/inventory/manager/recommender/recommendation_manager.py
@@ -4,19 +4,29 @@ import requests
 
 from bs4 import BeautifulSoup
 from spaceone.inventory.libs.manager import GoogleCloudManager
+from spaceone.inventory.connector.recommender.cloud_asset import CloudAssetConnector
 from spaceone.inventory.libs.schema.base import ReferenceModel
 from spaceone.inventory.model.recommender.recommendation.cloud_sevice_type import CLOUD_SERVICE_TYPES
 from spaceone.inventory.model.recommender.recommendation.cloud_service import RecommendationResource, \
     RecommendationResponse
 from spaceone.inventory.model.recommender.recommendation.data import Recommendation
-from spaceone.inventory.connector import RecommendationConnector
+from spaceone.inventory.connector import RecommendationConnector, InsightConnector
 
 _LOGGER = logging.getLogger(__name__)
+
+_RECOMMENDATION_TYPE_DOCS_URL = 'https://cloud.google.com/recommender/docs/recommenders'
+
+_UNAVAILABLE_RECOMMENDER_IDS = [
+    'google.cloudbilling.commitment.SpendBasedCommitmentRecommender',
+    'google.accounts.security.SecurityKeyRecommender',
+    'google.cloudfunctions.PerformanceRecommender'
+]
 
 
 class RecommendationManager(GoogleCloudManager):
     connector_name = 'RecommendationConnector'
     cloud_service_types = CLOUD_SERVICE_TYPES
+    project_id = None
 
     def collect_cloud_service(self, params):
         """
@@ -37,112 +47,144 @@ class RecommendationManager(GoogleCloudManager):
         error_responses = []
 
         secret_data = params['secret_data']
-        project_id = secret_data['project_id']
-        target_recommendation_names = params['recommendation_names']
+        self.project_id = secret_data['project_id']
 
-        recommendation_type_map = self._create_insight_type_by_crawling()
+        cloud_asset_conn: CloudAssetConnector = self.locator.get_connector(CloudAssetConnector, **params)
+        assets = cloud_asset_conn.list_assets_in_project()
+        asset_names = [asset['name'] for asset in assets]
+        target_locations = self._create_target_locations(asset_names)
+
+        recommender_id_map = self._create_recommendation_id_map_by_crawling()
+
+        recommendation_parents = self._create_recommendation_parents(target_locations, recommender_id_map)
+
         recommendation_conn: RecommendationConnector = self.locator.get_connector(RecommendationConnector, **params)
 
-        for recommendation_name in target_recommendation_names:
-            try:
-                recommendation = recommendation_conn.get_recommendation(recommendation_name)
-                recommendation_response = self._create_recommendation_response(recommendation, project_id,
-                                                                               recommendation_type_map)
-                collected_cloud_services.append(recommendation_response)
+        for recommendation_parent in recommendation_parents:
+            recommendations = recommendation_conn.list_recommendations(recommendation_parent)
+            for recommendation in recommendations:
+                try:
+                    region, recommender_id = self._get_region_and_recommender_id(recommendation['name'])
 
-            except Exception as e:
-                _LOGGER.error(f'[collect_cloud_service] => {e}', exc_info=True)
-                error_response = self.generate_resource_error_response(e, 'Recommender', 'Recommendation',
-                                                                       recommendation_name)
-                error_responses.append(error_response)
+                    display = {
+                        'recommender_id': recommender_id,
+                        'recommender_id_name': recommender_id_map[recommender_id]['name'],
+                        'recommender_id_description': recommender_id_map[recommender_id]['short_description'],
+                        'priority_display': self.convert_readable_priority(recommendation['priority']),
+                    }
 
-        collected_cloud_services.extend(self.cloud_service_types)
+                    if resource := recommendation['content']['overview'].get('resourceName'):
+                        display['resource'] = self._change_resource(resource)
+
+                    if insights := recommendation['associatedInsights']:
+                        insight_conn: InsightConnector = self.locator.get_connector(InsightConnector, **params)
+                        related_insights = self._list_insights(insights, insight_conn)
+                        display['insights'] = self._change_insights(related_insights)
+
+                    recommendation.update({
+                        'display': display
+                    })
+
+                    recommendation_data = Recommendation(recommendation, strict=False)
+                    recommendation_resource = RecommendationResource({
+                        'name': recommendation_data.get('name', 'Unknown'),
+                        'account': self.project_id,
+                        'tags': {},
+                        'region_code': region,
+                        'instance_type': '',
+                        'instance_size': 0,
+                        'data': recommendation_data,
+                        'reference': ReferenceModel(recommendation_data.reference())
+                    })
+
+                    recommendation_response = RecommendationResponse({'resource': recommendation_resource})
+                    collected_cloud_services.append(recommendation_response)
+
+                except Exception as e:
+                    _LOGGER.error(f'[collect_cloud_service] => {e}', exc_info=True)
+                    error_response = self.generate_resource_error_response(e, 'Recommender',
+                                                                           'Recommendation', recommendation)
+                    error_responses.append(error_response)
+
         _LOGGER.debug(f'** Recommender Recommendation Finished {time.time() - start_time} Seconds **')
         return collected_cloud_services, error_responses
 
     @staticmethod
-    def _create_insight_type_by_crawling():
-        res = requests.get("https://cloud.google.com/recommender/docs/recommenders")
+    def _create_recommendation_id_map_by_crawling():
+        res = requests.get(_RECOMMENDATION_TYPE_DOCS_URL)
         soup = BeautifulSoup(res.content, 'html.parser')
         table = soup.find("table")
         rows = table.find_all("tr")
 
-        recommendation_type_map = {}
+        recommendation_id_map = {}
         category = ''
         for row in rows:
             cols = row.find_all("td")
             cols = [ele.text.strip() for ele in cols]
             if cols:
                 try:
-                    category, name, recommendation_type, short_description = cols
+                    category, name, recommender_id, short_description = cols
                 except ValueError:
-                    name, recommendation_type, short_description = cols
+                    name, recommender_id, short_description = cols
 
-                if "\n" in recommendation_type:
-                    recommendation_type = recommendation_type.split("\n")
+                if recommender_id.count('google.') > 1:
+                    recommender_ids = []
+                    re_ids = recommender_id.split('google.')[1:]
+                    for re_id in re_ids:
+                        re_id = 'google.' + re_id
+                        if re_id not in _UNAVAILABLE_RECOMMENDER_IDS:
+                            recommender_ids.append(re_id)
                 else:
-                    recommendation_type = [recommendation_type]
+                    if recommender_id not in _UNAVAILABLE_RECOMMENDER_IDS:
+                        recommender_ids = [recommender_id]
+                    else:
+                        continue
 
-                for recommend_type in recommendation_type:
-                    recommendation_type_map[recommend_type] = {
+                for recommender_id in recommender_ids:
+                    recommendation_id_map[recommender_id] = {
                         'category': category,
                         'name': name,
                         'short_description': short_description
                     }
 
-        return recommendation_type_map
+        return recommendation_id_map
 
     @staticmethod
-    def _switch_insight_type_key_to_value(insight_type_map):
-        new_insight_type_map = {}
-        for key, value in insight_type_map.items():
-            if isinstance(value, list):
-                for element in value:
-                    new_insight_type_map[element] = key
-            else:
-                new_insight_type_map[value] = key
-        return new_insight_type_map
+    def _create_target_locations(asset_names):
+        locations = []
+        for asset_name in asset_names:
+            try:
+                prefix, sub_asset = asset_name.split('locations/')
+                location, _ = sub_asset.split('/', 1)
 
-    def _create_recommendation_response(self, recommendation, project_id, recommendation_type_map):
-        region, instance_type = self._get_region_and_instance_type(recommendation['name'])
+                if location not in locations:
+                    locations.append(location)
 
-        display = {
-            'instance_type': instance_type,
-            'instance_type_name': recommendation_type_map[instance_type]['name'],
-            'instance_type_description': recommendation_type_map[instance_type]['short_description'],
-            'priority_display': self.convert_readable_priority(recommendation['priority']),
-        }
+            except ValueError:
+                continue
 
-        if resource := recommendation['content']['overview'].get('resourceName'):
-            display['resource'] = self._change_resource(resource)
+        return locations
 
-        recommendation.update({
-            'display': display
-        })
+    def _create_recommendation_parents(self, locations, recommendation_type_map):
+        recommendation_parents = []
+        for location in locations:
+            for recommender_id in recommendation_type_map.keys():
+                recommendation_parents.append(
+                    f'projects/{self.project_id}/locations/{location}/recommenders/{recommender_id}'
+                )
 
-        recommendation_data = Recommendation(recommendation, strict=False)
-        recommendation_resource = RecommendationResource({
-            'name': recommendation_data.get('name', 'Unknown'),
-            'account': project_id,
-            'tags': {},
-            'region_code': region,
-            'instance_type': '',
-            'instance_size': 0,
-            'data': recommendation_data,
-            'reference': ReferenceModel(recommendation_data.reference())
-        })
-
-        return RecommendationResponse({'resource': recommendation_resource})
+        return recommendation_parents
 
     @staticmethod
-    def _get_region_and_instance_type(name):
+    def _get_region_and_recommender_id(recommendation_name):
         try:
-            project_id, resource = name.split('locations/')
+            project_id, resource = recommendation_name.split('locations/')
             region, _, instance_type, _ = resource.split('/', 3)
             return region, instance_type
 
         except Exception as e:
-            _LOGGER.error(f'[_get_region] recommendation passing error (data: {name}) => {e}', exc_info=True)
+            _LOGGER.error(f'[_get_region] recommendation passing error (data: {recommendation_name}) => {e}',
+                          exc_info=True)
 
     @staticmethod
     def convert_readable_priority(priority):
@@ -165,3 +207,43 @@ class RecommendationManager(GoogleCloudManager):
         except ValueError:
             return resource
 
+    @staticmethod
+    def _list_insights(insights, insight_conn):
+        related_insights = []
+        for insight in insights:
+            insight_name = insight['insight']
+            insight = insight_conn.get_insight(insight_name)
+            related_insights.append(insight)
+        return related_insights
+
+    def _change_insights(self, insights):
+        changed_insights = []
+        for insight in insights:
+            changed_insights.append({
+                'name': insight['name'],
+                'description': insight['description'],
+                'last_refresh_time': insight['lastRefreshTime'],
+                'observation_period': insight['observationPeriod'],
+                'state': insight['stateInfo']['state'],
+                'category': insight['category'],
+                'insight_subtype': insight['insightSubtype'],
+                'severity': insight['severity'],
+                'etag': insight['etag'],
+                'target_resources': self._change_target_resources(insight['targetResources'])
+            })
+        return changed_insights
+
+    def _change_target_resources(self, resources):
+        new_target_resources = []
+        for resource in resources:
+            new_target_resources.append({'name': resource,
+                                         'display_name': self._change_resource_name(resource)})
+        return new_target_resources
+
+    @staticmethod
+    def _change_resource_name(resource):
+        try:
+            resource_name = resource.split('/')[-1]
+            return resource_name
+        except ValueError:
+            return resource
