+++ src/spaceone/inventory/model/__init__.py
@@ -11,5 +11,4 @@ from spaceone.inventory.model.networking.route import *
 from spaceone.inventory.model.compute_engine.snapshot import *
 from spaceone.inventory.model.cloud_storage.bucket import *
 from spaceone.inventory.model.networking.vpc_network import *
-from spaceone.inventory.model.recommender.insight import *
 from spaceone.inventory.model.recommender.recommendation import *
