+++ src/spaceone/inventory/model/recommender/__init__.py
@@ -1,2 +1,1 @@
 from spaceone.inventory.model.recommender.recommendation import *
-from spaceone.inventory.model.recommender.insight import *
