--- src/spaceone/inventory/model/recommender/insight/widget/count_by_project.yml
@@ -1,15 +1,0 @@
----
-cloud_service_group: "Recommender"
-cloud_service_type: Insight
-name: Count by Project
-query:
-  aggregate:
-    - group:
-        keys:
-          - name: name
-            key: account
-        fields:
-          - name: value
-            operator: count
-options:
-  chart_type: DONUT
