--- src/spaceone/inventory/model/recommender/insight/widget/count_by_region.yml
@@ -1,20 +1,0 @@
----
-cloud_service_group: "Recommender"
-cloud_service_type: Insight
-name: Count by Region
-query:
-  aggregate:
-    - group:
-        keys:
-          - name: name
-            key: region_code
-        fields:
-          - name: value
-            operator: count
-options:
-  chart_type: COLUMN
-  name_options:
-    key: name
-    reference:
-      resource_type: "inventory.Region"
-      reference_key: region_code
