--- src/spaceone/inventory/model/recommender/insight/widget/total_count.yml
@@ -1,15 +1,0 @@
----
-cloud_service_group: "Recommender"
-cloud_service_type: Insight
-name: Total Count
-query:
-  aggregate:
-    - group:
-        fields:
-          - name: value
-            operator: count
-options:
-  value_options:
-    key: value
-    options:
-      default: 0
