+++ test/api/test_cloudservice_api.py
@@ -55,7 +55,7 @@ class TestCollector(TestCase):
         '''
         options = {
             # "cloud_service_types": ["CloudFunctions"],
-            "cloud_service_types": ["Recommender_v2"]
+            "cloud_service_types": ["Recommender"]
             # "custom_asset_url": 'http://xxxxx.spaceone.dev/icon/google'
         }
         filter = {}
